rename to architectures
